requirements.txt
flask==3.0.3
flask-cors==4.0.1
reportlab==4.2.2
python-dotenv==1.0.1
gunicorn==22.0.0
